+++ pkg/lang/csharp/asp_dotnet_core.go
@@ -94,19 +94,3 @@ func getDotnetCoreStartupClass(classNode *sitter.Node) (ASPDotNetCoreStartupClas
 
 	return startupClass, true
 }
-
-func FindSubtypes(unit *core.ExecutionUnit, baseNamespace, baseType string) []*TypeDeclaration {
-	var declarations []*TypeDeclaration
-	for _, csFile := range unit.FilesOfLang(CSharp) {
-		types := FindDeclarationsInFile[*TypeDeclaration](csFile).Declarations()
-		for _, t := range types {
-			for _, bNode := range t.Bases {
-				if IsValidTypeName(bNode, baseNamespace, baseType) {
-					declarations = append(declarations, t)
-					break
-				}
-			}
-		}
-	}
-	return declarations
-}
