--- pkg/lang/csharp/asp_dotnet_core_test.go
@@ -1,1 +1,0 @@
-package csharp
